fix:service start bug

+++ pay-service/pay-api/src/main/java/com/xcxcxcxcx/myshop/dto/PaymentRequest.java
@@ -19,7 +19,7 @@ public class PaymentRequest extends AbstractRequest{
 
     private Long userId;
 
-    private int payAmount;
+    private Double payAmount;
 
     private String description;
 
@@ -57,11 +57,11 @@ public class PaymentRequest extends AbstractRequest{
         this.userId = userId;
     }
 
-    public int getPayAmount() {
+    public Double getPayAmount() {
         return payAmount;
     }
 
-    public void setPayAmount(int payAmount) {
+    public void setPayAmount(Double payAmount) {
         this.payAmount = payAmount;
     }
